Change highlight settings; use ycm to lint rust
The new version of ycm for rust provides linting information using rls,
so ale is not needed. Set the color of ycm error so that it adheres to
ale's. I have never used java and am not going to use it in the
foreseeable future, so delete it as well.